Made to hydra ring variables int instead of double. Related to Issue [WTI-559]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@5705 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: b222ab4ab09d877a97fd6fcc484b4ac8232dab9d
Former-commit-id: 45446a80c47d0ee4298512b56aa78c3a81bc39e0

+++ Ringtoets/HydraRing/src/Ringtoets.HydraRing.Calculation/Data/Input/Structures/StructuresClosureCalculationInput.cs
@@ -35,7 +35,7 @@ namespace Ringtoets.HydraRing.Calculation.Data.Input.Structures
         private readonly double factorStormDurationOpenStructure;
         private readonly double failureProbabilityOpenStructure;
         private readonly double failureProbabilityReparation;
-        private readonly double identicalApertures;
+        private readonly int identicalApertures;
         private readonly double allowedLevelIncreaseStorageMean;
         private readonly double allowedLevelIncreaseStorageStandardDeviation;
         private readonly double modelFactorStorageVolumeMean;
@@ -86,7 +86,7 @@ namespace Ringtoets.HydraRing.Calculation.Data.Input.Structures
                                                     double factorStormDurationOpenStructure,
                                                     double failureProbabilityOpenStructure,
                                                     double failureProbabilityReparation,
-                                                    double identicalApertures,
+                                                    int identicalApertures,
                                                     double allowedLevelIncreaseStorageMean, double allowedLevelIncreaseStorageStandardDeviation,
                                                     double modelFactorStorageVolumeMean, double modelFactorStorageVolumeStandardDeviation,
                                                     double storageStructureAreaMean, double storageStructureAreaVariation,
